Move FileInfoFactory to statestore test code
